requirements-test.txt
django-nose
coverage
mock
responses
django-coverage
django-setuptest